Move dump substate under info
